==> 01_Extract_HDF5_Files_to_a_CSV_File.ipynb <==
# Extract information from HDF5 files and export to .csv file

## This dataset and extraction file of the HDF5 files, are courtesy of the Million Song Dataset created for machine learning.  For more information, please visit their site at https://labrosa.ee.columbia.edu/millionsong/

Thierry Bertin-Mahieux, Daniel P.W. Ellis, Brian Whitman, and Paul Lamere. 
The Million Song Dataset. In Proceedings of the 12th International Society
for Music Information Retrieval Conference (ISMIR 2011), 2011.

import os
import sys
import time
import glob
import tables
import scipy.io as sio
import pandas as pd
import numpy as np
import hdf5_getters
from pandas import read_hdf
import numpy as np

def get_all_files(basedir,ext='.h5') :
    """
    From a root directory, go through all subdirectories
    and find all files with the given extension.
    Return all absolute paths in a list.
    """
    allfiles = []
    for root, dirs, files in os.walk(basedir):
        files = glob.glob(os.path.join(root,'*'+ext))
        for f in files :
            allfiles.append( os.path.abspath(f) )
    return allfiles

#this uses the hdff_get_headers.py file to extract the data from the MSD.  Takes about 15 minutes to run

def transfer(h5path,matpath=None,force=False):
    """
    Transfer an HDF5 song file (.h5) to a matfile (.mat)
    If there are more than one song in the HDF5 file, each
    field name gets a number happened: 1, 2, 3, ...., numfiles
    PARAM
        h5path  - path to the HDF5 song file
        matpath - path to the new matfile, same as HDF5 path
                  with a different extension by default
        force   - if True and matfile exists, overwrite
    RETURN
        True if the file was transfered, False if there was
        a problem.
        Could also raise an IOException
    NOTE
        All the data has to be loaded in memory! be careful
        if one file contains tons of songs!
    """
    # sanity checks
    if not os.path.isfile(h5path):
        print ('path to HF5 files does not exist:',h5path)
        return False
    if not os.path.splitext(h5path)[1] == '.h5':
        print ('expecting a .h5 extension for file:',h5path)
        return False
    # check matfile
    if matpath is None:
        matpath = os.path.splitext(h5path)[0] + '.mat'
    if os.path.exists(matpath):
        if not force: 
            print('matfile',matpath,'already exists (delete or force):')
            return False
    # get all getters! we assume that all we need is in hdf5_getters.py
    # further assume that they have the form get_xxxxxx and that's the
    # only thing that has that form
    getters = list(filter(lambda x: x[:4] == 'get_', hdf5_getters.__dict__.keys()))
    getters.remove("get_num_songs") # special case
    # open h5 file
    h5 = hdf5_getters.open_h5_file_read(h5path)
    # transfer
    nSongs = hdf5_getters.get_num_songs(h5)
    matdata = {'transfer_note':'transferred on '+time.ctime()+' from file: '+h5path}
    try:
        # iterate over songs
        for songidx in range(nSongs):
            # iterate over getter
            for getter in getters:
                gettername = getter[4:]
                if nSongs > 1:
                    gettername += str(songidx+1)
                data = hdf5_getters.__getattribute__(getter)(h5,songidx)
                matdata[gettername] = data
    except MemoryError:
        print('Error')
        raise
    finally:
        # close h5
        h5.close()
    # create
    return matdata

h5s = get_all_files('/Users/samlundberg/desktop/capstonems/MillionSongSubset/data/')

df = pd.DataFrame()

%%time
for file in h5s:
    xd = transfer(file)
    df = df.append(pd.Series(xd), ignore_index=True)

dfcopy=df.copy()

#df=dfcopy.copy()

## Before saving to a .csv where the formatting changes, all of the song attrtibutes need to be converted from a list of values, to a mean of those values for further use in the modeling process later on

#this is taking the list of numbers for each audio measurement and finds the average to use later on
#the proper way would be to create a separate field for each value, but that would require a great deal more time 
#and subject matter expertese

#df['artist_terms_1'] = df['artist_terms'].apply(lambda col: np.mean(col))
df['artist_terms_freq_mean'] = df['artist_terms_freq'].apply(lambda col: np.mean(col))
df['artist_terms_weight_mean'] = df['artist_terms_weight'].apply(lambda col: np.mean(col))
df['segments_start_mean'] = df['segments_start'].apply(lambda col: np.mean(col))
df['segments_pitches_mean'] = df['segments_pitches'].apply(lambda col: np.mean(col))
df['segments_timbre_mean'] = df['segments_timbre'].apply(lambda col: np.mean(col))
df['segments_loudness_max_mean'] = df['segments_loudness_max'].apply(lambda col: np.mean(col))
df['segments_loudness_max_time_mean'] = df['segments_loudness_max_time'].apply(lambda col: np.mean(col))
df['segments_loudness_start_mean'] = df['segments_loudness_start'].apply(lambda col: np.mean(col))
df['sections_start_mean'] = df['sections_start'].apply(lambda col: np.mean(col))
df['sections_confidence_mean'] = df['sections_confidence'].apply(lambda col: np.mean(col))
df['beats_start_mean'] = df['beats_start'].apply(lambda col: np.mean(col))
df['beats_confidence_mean'] = df['beats_confidence'].apply(lambda col: np.mean(col))
df['bars_start_mean'] = df['bars_start'].apply(lambda col: np.mean(col))
df['bars_confidence_mean'] = df['bars_confidence'].apply(lambda col: np.mean(col))
df['tatums_start_mean'] = df['tatums_start'].apply(lambda col: np.mean(col))
df['tatums_confidence_mean'] = df['tatums_confidence'].apply(lambda col: np.mean(col))

df.head()

#set paramater values
a,b,c,d=[df['artist_name'],df['artist_terms'] ,df['artist_terms_freq'],df['artist_terms_weight']]

a[0],b[0][0],c[0][0],d[0][0]

a[0],b[0][1],c[0][1],d[0][1]

a[0][1],b[0][1],c[0][1]

a[1],b[1][0],c[1][0],d[1][0]

b[1]

#drop rows with a blank list [] in artist_terms so there will be a genre label for every artist
df=df[df['artist_terms'].map(lambda i: len(i)) > 0]

#find list of genres 
i=0
list_g=[]

acount=0
count=0
for j in range((a.shape[0])):
    for i in range((b[j].shape[0])):
        if len(b[j]) != 0:
        #print(b[i][0])

            genre=b[j][i]
            list_g.append(genre)
            count=(count + 1)
            i= count

    #print(list_g)        


#there are a huge number of genres being tracked, over 3500
#list_g

#create unique values of genre (remove duplicates)
g=[el for i, el in enumerate(list_g) if el not in list_g[:i]]

#this creates a genre column in the df and uses the top genre value as the defining genre for that artist

df['genre']=""
list_a=[]

j=0
for j in range((a.shape[0])):
    if len(b[j]) != 0:
        i=0

        #for i in range((b[j].shape[0])):
        #all=[a[j],b[j][0]]
        
        #all=b[j][0]
        #list_a.append(all)

        #print(all)
        #df['genre']=pd.Series(list_a)
        df['genre'][j]=b[j][0]

#reset index numbers
df.reset_index(drop=True, inplace=True)

df.head()

#create decade column

df['year']=df['year'].astype(int)

# NEAREST 10 FOR DECADE START
df['decade_start'] = (df['year'] // 10) * 10 + 1

# ADJUST FOR YEARS ENDING IN ZERO
df.loc[(df['year'] % 10) == 0, 'decade_start'] = df['decade_start'] - 10

# CALCULATE DECADE RANGE
#df['decade'] = df['decade_start'].astype('str') + ' - ' + \
#                     (df['decade_start'] + 9).astype('str') 

df['decade'] = df['decade_start'].astype('int') -1

df.drop(['decade_start'], inplace=True, axis=1)


## Create genre table that lists all genres, and their related scores, for each artist

#del genre_alt
#create unique values of genre (remove duplicates)
g=[el for i, el in enumerate(list_g) if el not in list_g[:i]]

#create genre df with column labels as genre type and empty df
genre_alt=pd.DataFrame(g).T
genre_alt.columns=g
genre_alt=genre_alt.drop(genre_alt.index[[0]])
genre_alt.insert(loc=0, column='artist', value="")

#this is good-takes quite awhile to run
#attempt to just pull just first genre, score from each artist

list_a=[]
j=0

for j in range((a.shape[0])):
    if len(b[j]) != 0:
        i=0

        #all=[a[j],b[j][0],c[j][0] * d[j][0]]
        genre_val=c[j][0] * d[j][0]
        list_a.append(all)

        genre_dict=({'artist':a[j],b[j][i]:genre_val})
        genre1  = pd.DataFrame(genre_dict, index=[j])
        genre_alt = pd.concat([genre_alt, genre1], sort=False)


genre_alt.fillna(0, inplace=True)

genre_alt.drop_duplicates(['artist'], inplace=True)
genre_alt = genre_alt.reset_index()

genre_alt

#alt method to populate genre table with values-not as efficient

#del genre
#create unique values of genre (remove duplicates)
g=[el for i, el in enumerate(list_g) if el not in list_g[:i]]

#create genre df with column labels as genre type and empty df
genre=pd.DataFrame(g).T
genre.columns=g
genre=genre.drop(genre.index[[0]])
genre.insert(loc=0, column='artist', value="")

#genrecopy=genre.copy()
#genre=genrecopy.copy()

test=df.copy()

#set a,b,c,d values for loop below to find genres and respecitve values
a,b,c,d=[test['artist_name'],test['artist_terms'] ,test['artist_terms_freq'],test['artist_terms_weight']]

a[0][0],b[0][0],c[0][0],d[0][0]

#creates master genre table for each genre for each artist
#takes a LONG time to run

list_gen={}
art_dict={}


for j in range((a.shape[0])):
    list_gen={}
    art_dict={}
    art_dict={'artist':a[j]}

    for i in range((b[j].shape[0])):
        genre_val=c[j][i] * d[j][i]
        genre_dict={b[j][i]:genre_val}
        list_gen.update(genre_dict)

    art_dict.update(list_gen)
    genre1  = pd.DataFrame(art_dict, index=[j])
    genre = pd.concat([genre, genre1], sort=False)      

#fill NaN values with 0
genre.fillna(0, inplace=True)

genre.head()

#drop duplicate rows of artist genre values
genre.drop_duplicates(['artist'], inplace=True)
genre = genre.reset_index()

#export df to csv file
df.to_csv('./data/million_songs.csv')

#export genre table to csv
genre.to_csv('./data/genre.csv')

#export genre_alt table to csv
genre_alt.to_csv('./data/genrealt.csv')
