==> 03_EDA.ipynb <==
# Exploratory Data Analysis of original MSD dataset

import pandas as pd
import numpy as np
import requests, base64, json, logging
from sklearn.feature_extraction.text import TfidfVectorizer, CountVectorizer

df=pd.read_csv('./Data/MILLION_SONGS.csv')

df.head()

df.info()

#drop columns not needed
df.drop(['Unnamed: 0','artist_7digitalid',
         'track_7digitalid','transfer_note','artist_mbid', 'artist_mbtags', 'artist_mbtags_count', 
         'artist_playmeid', 'audio_md5','analysis_sample_rate','transfer_note','release_7digitalid',
         'segments_start','segments_pitches','segments_timbre',
         'segments_loudness_max','segments_loudness_max_time','segments_loudness_start','sections_start',
         'sections_confidence','beats_start','beats_confidence','bars_start','bars_confidence','tatums_start',
         'tatums_confidence'], axis=1, inplace=True)
#'artist_terms_freq','artist_terms_weight',

#replace strings that start with b'
#remove other punctuation

df['artist_id']=df['artist_id'].str.replace("b'","")
df['artist_name']=df['artist_name'].str.replace("b'","")
df['artist_terms']=df['artist_terms'].str.replace("b'","")
df['song_id']=df['song_id'].str.replace("b'","")
df['title']=df['title'].str.replace("b'","")
df['track_id']=df['track_id'].str.replace("b'","")
df['release']=df['release'].str.replace("b'","")
df['similar_artists']=df['similar_artists'].str.replace("b'","")
df['genre']=df['genre'].str.replace("b'","")
df['artist_location']=df['artist_location'].str.replace("b'","")

df['artist_id']=df['artist_id'].str.replace('b"',"")
df['artist_name']=df['artist_name'].str.replace('b"',"")
df['artist_terms']=df['artist_terms'].str.replace('b"',"")
df['song_id']=df['song_id'].str.replace('b"',"")
df['title']=df['title'].str.replace('b"',"")
df['track_id']=df['track_id'].str.replace('b"',"")
df['release']=df['release'].str.replace('b"',"")
df['similar_artists']=df['similar_artists'].str.replace('b"',"")
df['genre']=df['genre'].str.replace('b"',"")
df['artist_location']=df['artist_location'].str.replace('b"',"")

df['artist_id']=df['artist_id'].str.replace("'","")
#df['artist_name']=df['artist_name'].str[:-1]
df['artist_terms']=df['artist_terms'].str.replace("'","")
df['song_id']=df['song_id'].str.replace("'","")
df['title']=df['title'].str.replace("'","")
df['title']=df['title'].str.replace('"',"")
df['track_id']=df['track_id'].str.replace("'","")
#df['artist_location']=df['artist_location'].str[:-1]

#df['artist_terms']=df['artist_terms'].str.replace("'","")
df['similar_artists']=df['similar_artists'].str.replace("'","")
df['artist_name']=df['artist_name'].str.replace("'","")
df['genre']=df['genre'].str.replace("'","")
df['release']=df['release'].str.replace("'","")
df['artist_location']=df['artist_location'].str.replace("'","")

#year column is missing 5315 values, or over 50% of dataset
df['year'].value_counts()

#look at fields that have missing values
#artist location is missing many values, not necessarily a good or bad thing.  Will leave as is for now and handle
#on a case by case basis in the modeling stage
#song hotttnesss is missing a number of values, which could be a problem in predicting if a song is hot or not.  
#will handle this at that stage
df.isnull().sum().sort_values(ascending=False) 

#only 4 missing values from artist_familiarity
#none of the artists exist in the dataset with a value, will fill with median familiarity from all artists
df.loc[df['artist_familiarity'].isnull()]

df["artist_familiarity"] = df["artist_familiarity"].fillna(df["artist_familiarity"].median())

#the mode, or key type (major/minor) of a song is important to the prediction of popularity.  This is currently
#scaled as 0 for minor, 1 for major keys.  Looking at the confidence % of each of these, it makes more sense to 
#change minor keys to -1 to 0 and major to 0 to 1 as the values in the df.  -1 is 100% minor, +1 is 100% major

#df = df.drop("song_hotttnesss", axis=1)
df['mode_orig']=df['mode']

for i, row in df.iterrows():
    if df['mode'][i] == 1:
        df['mode'][i] = df['mode_confidence'][i]
    else:
        df['mode'][i] = -df['mode_confidence'][i]
#df = df.drop('mode_confidence', axis=1)

## Add Billboard Top 100 .csv file to DF and create field BB_Hot to determine if a song was popular or not for later modeling

billboard = pd.read_csv("./Data/billboard.csv")
billboard.head()

#Creating a function that takes a list of the track names from the BillBoard top 100 that was
#webscraped and look for a matching title in the MSD subset.  If it finds a matching title, this
#means it was a popular song, and creating a binary column 1=True, 0=False to track this for modeling purposes

billboard_tracks = list(billboard.track)

def match_title(row):
    bb_list = billboard_tracks
    if row['title'] in bb_list:
        return 1
    else:
        return 0

#add to df
df['bb_hot'] = df.apply(match_title, axis=1)

#see how many bb_hot 100 titles exist in the MSD
df['bb_hot'].value_counts()

#% of hot or not ~10.6%
len(df[df['bb_hot']==1])/len(df['bb_hot'])

df[df['bb_hot']==1]

#export as csv file
df.to_csv("./Data/datacleanedwithbb.csv")
